requirements.txt
# Dependências do projeto

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data Science & Finance
pandas>=2.1.0
numpy>=1.24.0
yfinance>=0.2.28
scipy>=1.11.0

# AI/ML
groq>=0.4.0

# Environment Variables
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.25.0

# Database
psycopg2-binary>=2.9.9

# Security
cryptography>=41.0.0
